Split out EPGs into separate classes with new names. Add new optional skin names. Open grid EPG in correct place when EPG history is large. Fix grid EPG selection artefact bug when empty cells are visible. Fix grid EPG disappearing selection bug. Removed confusing bouquet selection from help. EPG preview mode now restarts playback of recording. Fix multi EPG to remember position when adding timers. Allow more EPG history to be kept. Reduce HD/FHD code duplication.
