+++ data/menu.xml
@@ -78,11 +78,11 @@ self.session.openWithCallback(msgClosed, EpgDeleteMsg)
 </code>
 					</item>
 				</menu>
+				<item level="2" entryID="setup_epgsingle"><setup id="epgsingle"/></item>
+				<item level="2" entryID="setup_epggrid"><setup id="epggrid"/></item>
+				<item level="2" entryID="setup_epginfobarsingle"><setup id="epginfobarsingle"/></item>
+				<item level="2" entryID="setup_epginfobargrid"><setup id="epginfobargrid"/></item>
 				<item level="2" entryID="setup_epgmulti"><setup id="epgmulti"/></item>
-				<item level="2" entryID="setup_epgenhanced"><setup id="epgenhanced"/></item>
-				<item level="2" entryID="setup_epginfobar"><setup id="epginfobar"/></item>
-				<item level="2" entryID="setup_epginfobargraphical"><setup id="epginfobargraphical"/></item>
-				<item level="2" entryID="setup_epggraphical"><setup id="epggraphical"/></item>
 			</menu>
 			<!-- Menu / Setup / User Interface -->
 			<menu weight="4" text="User interface" entryID="osd_setup" requires="OsdMenu">
