+++ lib/python/Tools/TextBoundary.py
@@ -2,6 +2,6 @@ from enigma import eLabel
 
 # Calls onto the static function in eLabel. This avoids causing an invalidate
 # on the parent container which is detrimental to UI performance, 
-# particularly in a complex screen like the graph EPG
+# particularly in a complex screen like the grid EPG
 def getTextBoundarySize(instance, font, targetSize, text):
 	return eLabel.calculateTextSize(font, text, targetSize)
